v0.9.4.0 Features de troca de produto adicionados

+++ EasyPDV/UI/FrmReversal.cs
@@ -1,9 +1,9 @@
-﻿using EasyPDV.Model;
+﻿using EasyPDV.Entities;
+using EasyPDV.Model;
 using System;
-using System.Windows.Forms;
 using System.Collections.Generic;
-using EasyPDV.Entities;
 using System.Linq;
+using System.Windows.Forms;
 
 namespace EasyPDV.UI
 {
@@ -13,6 +13,13 @@ namespace EasyPDV.UI
         IndividualSaleDAO individualSaleDAO = new IndividualSaleDAO();
         IndividualSale individualSale = new IndividualSale();
         Product product = new Product();
+        CashierOpenDAO cashierOpenDAO = new CashierOpenDAO();
+        ReversedSale changeSale = new ReversedSale();
+        ReversedSaleDAO changeSaleDAO = new ReversedSaleDAO();
+        string[] splitProduct;
+        string[] splitProductChange;
+        double FromProduct;
+        double ToProduct;
         public FrmReversal()
         {
             InitializeComponent();
@@ -27,11 +34,13 @@ namespace EasyPDV.UI
         {
             lblChange.Visible = false;
             comboProductChange.Visible = false;
+            ToChangeProductsCount.Visible = false;
         }
         public void UnhideStuff()
         {
             lblChange.Visible = true;
             comboProductChange.Visible = true;
+            ToChangeProductsCount.Visible = true;
         }
 
         public void PopulateCombos()
@@ -43,11 +52,11 @@ namespace EasyPDV.UI
             List<Product> noRepeatSoldProducts = soldProducts.Distinct().ToList();
             foreach (Product product in noRepeatSoldProducts)
             {
-                comboProduct.Items.Add(product.Name + " | " + productDAO.GetPrice(product));
+                comboProduct.Items.Add(product.Name + " | " + productDAO.GetPrice(product.Name));
             }
             foreach (Product product in products)
             {
-                comboProductChange.Items.Add(product.Name + " | " + productDAO.GetPrice(product));
+                comboProductChange.Items.Add(product.Name + " | " + productDAO.GetPrice(product.Name));
             }
         }
 
@@ -68,22 +77,56 @@ namespace EasyPDV.UI
             DialogResult dialogResult = MessageBox.Show("Confirma ação?", "Realizar Troca/Estorno", MessageBoxButtons.OKCancel);
             if (actionCombo.Text == "Trocar")
             {
-                if (dialogResult == DialogResult.OK && comboProductChange.Text != "" && comboProduct.Text != "")
+                if (SoldProductsCount.Value > 0 && ToChangeProductsCount.Value > 0)
                 {
-                    string[] splitProductChange = comboProductChange.Text.Split('|');
-                    string[] splitProduct = comboProduct.Text.Split('|');
-                    individualSale.Product = splitProductChange[0].Trim();
-                    individualSale.SaleDate = DateTime.Now;
-                    individualSale.SalePrice = double.Parse(splitProductChange[1].Trim());
-                    individualSale.PaymentMethod = "Troca ou estorno";
-                    individualSaleDAO.DeleteIndividualSale(splitProduct[0].Trim());
-                    individualSaleDAO.InsertIndividualSale(individualSale);
-                    product.Name = splitProduct[0].Trim();
-                    productDAO.SumStock(product);
-                    product.Name = splitProductChange[0].Trim();
-                    product.ID = productDAO.GetID(product);
-                    productDAO.SubtractStock(product);
+                    if (dialogResult == DialogResult.OK && comboProductChange.Text != "" && comboProduct.Text != "")
+                    {
+
+                        for (int i = 0; i < SoldProductsCount.Value; i++)
+                        {
+
+                            individualSaleDAO.DeleteIndividualSale(splitProduct[0].Trim());
+                            product.Name = splitProduct[0].Trim();
+                            changeSale.ProductChangeFrom = splitProduct[0].Trim() + " x" + SoldProductsCount.Value.ToString();
+                            productDAO.SumStock(product);
+
+                        }
+
+                        for (int i = 0; i < ToChangeProductsCount.Value; i++)
+                        {
+
+                            product.Name = splitProductChange[0].Trim();
+                            changeSale.ProductChangeTo = splitProductChange[0].Trim() + " x" + ToChangeProductsCount.Value.ToString();
+                            product.ID = productDAO.GetID(product);
+                            productDAO.SubtractStock(product);
+                            individualSale.Product = splitProductChange[0].Trim();
+                            individualSale.SaleDate = DateTime.Now;
+                            individualSale.SalePrice = double.Parse(splitProductChange[1].Trim());
+                            individualSale.PaymentMethod = "Troca ou estorno";
+                            individualSaleDAO.InsertIndividualSale(individualSale);
+                            RawPrinterHelper.Print(
+                                   "-------------------\n\n" +
+                                   TypeHelper.FormatToCenter(cashierOpenDAO.ReturnEventName().Trim()) + "\n" +
+                                   TypeHelper.FormatToCenter(DateTime.Now.ToString("d").Trim()) + "\n" +
+                                   "\n" + TypeHelper.FormatToCenter(product.Name.ToUpper().Trim()) + "\n\n" +
+                                   "-------------------"
+                                   );
+                        }
+
+                        changeSale.ChangeType = "Troca";
+                        changeSale.SaleDate = DateTime.Now;
+                        FromProduct = double.Parse(splitProduct[1]) * (int)SoldProductsCount.Value;
+                        ToProduct = double.Parse(splitProductChange[1]) * (int)ToChangeProductsCount.Value;
+                        changeSale.Balance = ToProduct - FromProduct;
+                        changeSaleDAO.InsertChangedSale(changeSale);
+
+                        comboProduct.Items.Clear();
+                        comboProductChange.Items.Clear();
+                        PopulateCombos();
+
+                    }
                 }
+
                 else
                 {
                     MessageBox.Show("Preencha todos os campos corretamente!", "Aviso", MessageBoxButtons.OK, MessageBoxIcon.Warning);
@@ -91,30 +134,51 @@ namespace EasyPDV.UI
             }
             if (actionCombo.Text == "Estornar")
             {
-                if (dialogResult == DialogResult.OK && comboProduct.Text != "")
+                if (SoldProductsCount.Value > 0)
                 {
-                    string[] splitProduct = comboProduct.Text.Split('|');
-                    individualSaleDAO.DeleteIndividualSale(splitProduct[0].Trim());
-                    product.Name = splitProduct[0].Trim();
-                    productDAO.SumStock(product);
+                    if (dialogResult == DialogResult.OK && comboProduct.Text != "")
+                    {
+                        for (int i = 0; i < SoldProductsCount.Value; i++)
+                        {
+                            changeSale.ProductChangeFrom = splitProduct[0].Trim() + " x" + SoldProductsCount.Value.ToString();
+                            individualSaleDAO.DeleteIndividualSale(splitProduct[0].Trim());
+                            product.Name = splitProduct[0].Trim();
+                            productDAO.SumStock(product);
+                        }
+
+                        changeSale.ProductChangeTo = " - ";
+                        changeSale.ChangeType = "Estorno";
+                        changeSale.SaleDate = DateTime.Now;
+                        FromProduct = double.Parse(splitProduct[1]) * (int)SoldProductsCount.Value;
+                        changeSale.Balance = FromProduct * -1;
+                        changeSaleDAO.InsertChangedSale(changeSale);
+
+                        comboProduct.Items.Clear();
+                        comboProductChange.Items.Clear();
+                        PopulateCombos();
+                    }
                 }
+
                 else
                 {
-                    MessageBox.Show("Preencha o item a trocar", "Aviso", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                    MessageBox.Show("Preencha todos os campos corretamente!", "Aviso", MessageBoxButtons.OK, MessageBoxIcon.Warning);
                 }
             }
-            comboProduct.Items.Clear();
-            comboProductChange.Items.Clear();
-            PopulateCombos();
+
         }
 
         private void comboProduct_SelectedIndexChanged(object sender, EventArgs e)
         {
 
+            splitProduct = comboProduct.Text.Split('|'); // product sold
+            SoldProductsCount.Maximum = individualSaleDAO.ReadTotalIndividualSoldProduct(splitProduct[0].Trim());
         }
 
         private void comboProductChange_SelectedIndexChanged(object sender, EventArgs e)
         {
+            splitProductChange = comboProductChange.Text.Split('|'); // product to give in change
+            product.Name = splitProductChange[0].Trim();
+            ToChangeProductsCount.Maximum = productDAO.CheckStock(product);
 
         }
     }
